+++ tests/unit/station/test_station_device.py
@@ -1574,6 +1574,53 @@ def test_stations_wren_trl(station_device: SpsStation) -> None:
     assert station_device.WrenTRL == get_wren_name()
 
 
+def test_stations_wren_health_check_fail_on_timeout(station_device: SpsStation) -> None:
+    """
+    Test that SPSStation properly exposes WREN health check fail on timeout flag.
+
+    :param station_device: The station device to use.
+
+    """
+    # Check the initial value is the same as the device property
+    assert (
+        station_device.WrenHealthCheckFailOnTimeout
+        == station_device.WRENHealthCheckFailOnTimeout
+    )
+
+    # Toggle the value
+    new_value = not station_device.WRENHealthCheckFailOnTimeout
+
+    # Set the value
+    station_device.WrenHealthCheckFailOnTimeout = (  # type: ignore[method-assign]
+        new_value
+    )
+
+    # Check the new value has been set
+    assert station_device.WrenHealthCheckFailOnTimeout == new_value
+
+
+def test_stations_wren_health_check_timeout(station_device: SpsStation) -> None:
+    """
+    Test that SPSStation properly exposes WREN health check timeout.
+
+    :param station_device: The station device to use.
+
+    """
+    # Check the initial value is the same as the device property
+    assert (
+        station_device.WrenHealthCheckTimeout == station_device.WRENHealthCheckTimeout
+    )
+
+    # Increment the value
+    new_value = station_device.WRENHealthCheckTimeout + 1
+
+    # Set the value
+    station_device.WrenHealthCheckTimeout = new_value  # type: ignore[method-assign]
+
+    # Check the new value has been set
+    assert station_device.WrenHealthCheckTimeout == new_value
+
+
 def test_AcquireDataForCalibration(
     station_device: SpsStation,
     daq_device: DeviceProxy,
@@ -1912,16 +1959,21 @@ def test_programing_state_health_rollup(
                 "subrack_failed": 0.2,
                 "tile_degraded": 0.05,
                 "tile_failed": 0.2,
+                "wren_degraded": 1.0,
+                "wren_failed": 1.0,
                 "pps_delta_degraded": 4,
                 "pps_delta_failed": 9,
                 "subracks": [1, 1, 1],  # Expect these to be overwritten
                 "tiles": [1, 1, 2],  # Expect these to be overwritten
+                "wren": [1, 1, 1],
             },
             {
                 "subrack_degraded": 0.1,
                 "subrack_failed": 0.3,
                 "tile_degraded": 0.07,
                 "tile_failed": 0.2,
+                "wren_degraded": 0.5,
+                "wren_failed": 0.5,
                 "pps_delta_degraded": 6,
                 "pps_delta_failed": 10,
             },